move npm utils to utils.npm
