+++ src/utils.npm.ts
@@ -6,7 +6,7 @@ import execa from 'execa';
 import { pathExists } from './utils.file';
 import { startSpinner, setSpinner, failSpinner, succeedSpinner } from './utils.spinner';
 import { dirname } from 'path';
-import { NpmList, NpmListDependency } from '.';
+import { NpmList, PackageWithVersion } from '.';
 
 export const TYPE_DEFINITION_FILE_NAME = 'index.d.ts';
 export const TMP_FOLDER = '.tmp';
@@ -228,6 +228,60 @@ export async function getNpmPackageVersionFromProjectPath(path: string, pkgName:
       }
     }
   } catch (e) {
+    console.log('Could not retrieve version information about the package ' + pkgName);
     return;
   }
+}
+
+const packageWithVersionRe = /(.+)(@.*?$)/;
+
+/**
+ * Given a string representation of a package and version (e.g. "package@version")
+ * returns an array of objects wih the parsed package name and version.
+ */
+export async function resolveTargetPackages(target: string): Promise<PackageWithVersion[]> {
+  const packages: PackageWithVersion[] = [];
+  const items = target.trim().split(',');
+  for (const item of items) {
+    let pkg: PackageWithVersion;
+    const trimmedItem = item.replace(/\s/g, '');
+    if (trimmedItem === '') {
+      continue;
+    }
+    const match = packageWithVersionRe.exec(trimmedItem);
+    if (match) {
+      pkg = {
+        name: match[1].trim(),
+        version: match[2].substring(1).trim(),
+      };
+    } else {
+      pkg = {
+        name: item.trim(),
+        version: 'latest',
+      };
+    }
+    console.log(`🔍 Resolving ${pkg.name}@${pkg.version}...`);
+    const resolvedVersion = await resolvePackageVersion(pkg.name, pkg.version);
+    if (resolvedVersion) {
+      packages.push({ ...pkg, version: resolvedVersion });
+    } else {
+      throw new Error(`Could not find package ${pkg.name}@${pkg.version} in the npm registry`);
+    }
+  }
+
+  return packages;
+}
+
+/**
+ * Given a package and version it returns the resolved
+ * version of the package using the npm registry.
+ *
+ * If a package or its version does not exist it returns undefined.
+ */
+async function resolvePackageVersion(pkg: string, version = 'latest'): Promise<string | void> {
+  const details = await getNpmPackageDetails(pkg, version);
+  if (details && details.version !== '') {
+    return details.version;
+  }
+  return;
 }
